fix(bridge): move BridgeSettings to repositories layer per Aristotle review
