+++ bridge/app/lib/src/repositories/bridge_settings.dart
